+++ src/nvim/profile.h
@@ -5,6 +5,7 @@
 #include <time.h>
 
 #include "nvim/ex_cmds_defs.h"
+#include "nvim/runtime.h"
 
 #define TIME_MSG(s) do { \
   if (time_fd != NULL) time_msg(s, NULL); \
